Update Callback Function to Print Current State
